backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.8.0
httpx>=0.27.0
qdrant-client>=1.10.1
langdetect>=1.0.9
python-multipart>=0.0.9
SQLAlchemy>=2.0.30
alembic>=1.13.2
psycopg[binary]>=3.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
